internal/config: remove UseLegacyWorkflow option (#962)
This option was added to temporarily enable the legacy credential chain evaluation order from V1 of this library to be adopted within V2. As of Terraform 1.8, the ability to use this legacy behavior has been completely removed, and the credential chain is always evaluated in the same order as the AWS SDK. As this option is no longer in use by any consumers of this library, it can be safely removed here.